feat(#2148): Move classes into CaptureLogs
